feature : logout 기능 개발 완료. 테스트 필요
# Conflicts:
#	src/main/java/com/ndgl/spotfinder/domain/user/service/UserService.java
