+++ src/main/java/com/ndgl/spotfinder/domain/user/controller/UserController.java
@@ -12,8 +12,7 @@ import org.springframework.web.bind.annotation.RestController;
 import com.ndgl.spotfinder.domain.user.dto.UserJoinRequest;
 import com.ndgl.spotfinder.domain.user.dto.UserLoginResponse;
 import com.ndgl.spotfinder.domain.user.entity.Oauth;
-import com.ndgl.spotfinder.domain.user.service.UserJoinService;
-import com.ndgl.spotfinder.domain.user.service.UserLoginService;
+import com.ndgl.spotfinder.domain.user.service.OauthService;
 import com.ndgl.spotfinder.domain.user.service.UserService;
 import com.ndgl.spotfinder.global.exception.ErrorCode;
 import com.ndgl.spotfinder.global.rsdata.RsData;
@@ -26,15 +25,13 @@ import jakarta.validation.Valid;
 @RequestMapping("/api/v1/users")
 public class UserController {
 
-	private final UserLoginService userLoginService;
-	private final UserJoinService userJoinService;
+	private final OauthService oauthService;
 	private final TokenProvider tokenProvider;
 	private final UserService userService;
 
-	public UserController(UserLoginService userLoginService, UserJoinService userJoinService,
+	public UserController(OauthService oauthService,
 		TokenProvider tokenProvider, UserService userService) {
-		this.userLoginService = userLoginService;
-		this.userJoinService = userJoinService;
+		this.oauthService = oauthService;
 		this.tokenProvider = tokenProvider;
 		this.userService = userService;
 	}
@@ -45,7 +42,7 @@ public class UserController {
 	public RsData<Void> join(
 		@Valid @RequestBody UserJoinRequest userJoinRequest) {
 
-		userJoinService.join(userJoinRequest);
+		userService.join(userJoinRequest);
 
 		return RsData.success(HttpStatus.OK);
 	}
@@ -57,7 +54,7 @@ public class UserController {
 		HttpServletResponse response
 	) {
 		//  구글 로그인 처리
-		UserLoginResponse responseDto = userLoginService.processGoogleLogin(Oauth.Provider.GOOGLE, code, redirectUri,
+		UserLoginResponse responseDto = oauthService.processGoogleLogin(Oauth.Provider.GOOGLE, code, redirectUri,
 			response);
 
 		return new RsData<>(responseDto.getCode(), responseDto.getMessage(), responseDto);
@@ -69,11 +66,16 @@ public class UserController {
 		@CookieValue(value = "accessToken", required = false) String accessToken,
 		HttpServletResponse response
 	) {
+		//  accessToken 확인
 		if (accessToken == null) {
 			ErrorCode.MISSING_ACCESS_TOKEN.throwServiceException();
 		}
 
 		String userId = tokenProvider.getEmail(accessToken);
 
+		//  로그아웃 처리
+		userService.logout(userId, response, accessToken);
+
+		return RsData.success(HttpStatus.OK);
 	}
 }
